Refactored metadata_library
Moved library file loading into metadata_library.go to make it reusable for future tools like graphql schema generator.
