+++ internal/model/library/metadata_library.go
@@ -1,7 +1,13 @@
 package library
 
 import (
-	"github.com/opendatahub-io/model-registry/internal/ml_metadata/proto"
+	"fmt"
+	"github.com/golang/glog"
+	"gopkg.in/yaml.v3"
+	"io/fs"
+	"os"
+	"path/filepath"
+	"strings"
 )
 
 //go:generate go-enum -type=PropertyType
@@ -49,10 +55,46 @@ type MetadataLibrary struct {
 	ExecutionTypes []ExecutionType `yaml:"execution-types,omitempty"`
 }
 
-func ToProtoProperties(props map[string]PropertyType) map[string]proto.PropertyType {
-	result := make(map[string]proto.PropertyType)
-	for name, prop := range props {
-		result[name] = proto.PropertyType(prop)
+func LoadLibraries(dirs []string) (map[string]*MetadataLibrary, error) {
+	result := make(map[string]*MetadataLibrary)
+	for _, dir := range dirs {
+		abs, err := filepath.Abs(dir)
+		if err != nil {
+			return nil, fmt.Errorf("error getting absolute library path for %s: %w", dir, err)
+		}
+		_, err = os.Stat(abs)
+		if err != nil {
+			return nil, fmt.Errorf("error opening library path for %s: %w", abs, err)
+		}
+		err = filepath.WalkDir(abs, func(path string, entry fs.DirEntry, err error) error {
+			if err != nil {
+				glog.Warningf("error reading library path %s: %v", path, err)
+				return filepath.SkipDir
+			}
+			if entry.IsDir() || !isYamlFile(path) {
+				return nil
+			}
+
+			bytes, err := os.ReadFile(path)
+			if err != nil {
+				return fmt.Errorf("failed to read library file %s: %w", path, err)
+			}
+			lib := &MetadataLibrary{}
+			err = yaml.Unmarshal(bytes, lib)
+			if err != nil {
+				return fmt.Errorf("failed to parse library file %s: %w", path, err)
+			}
+			result[path] = lib
+			return nil
+		})
+		if err != nil {
+			return nil, fmt.Errorf("failed to read library directory %s: %w", abs, err)
+		}
 	}
-	return result
+	return result, nil
+}
+
+func isYamlFile(path string) bool {
+	lowerPath := strings.ToLower(filepath.Ext(path))
+	return strings.HasSuffix(lowerPath, ".yaml") || strings.HasSuffix(lowerPath, ".yml")
 }
